Uploading ScratchWallet project to github for community support.
A scratch wallet is a type of web3 wallet that contains Unknown Private Keys or UPKs. Once the UPK becomes exposed by scratching, the scratch wallet is destroyed and the wallet becomes a paper wallet. A UPK must only be stored on the scratch wallet, no other memory shall exist, otherwise it is not a UPK, just a Private Key. In this way we can give crypto tokens physical only representation. A token in a wallet with a UPK  has essentially been withdrawn off the blockchain as it can't be transacted or used until you scratch the UPK. A UPK could be a key below a scratch surface ot it could be encased in a breakable material, like porcelain so the physical object must be destroyed for it to be used as a